+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- `jbpy.core.DataExtensionSegment.load()` for DES subheaders with conditional, user-defined fields
+
 
 ## [0.5.0] - 2025-12-16